fix issue with tests
